add getter for ui

+++ contracts/PerpdexMarket.sol
@@ -368,7 +368,7 @@ contract PerpdexMarket is IPerpdexMarket, ReentrancyGuard, Ownable, Multicall {
         external
         view
         returns (
-            uint256 executionId,
+            uint48 executionId,
             uint256 executedBase,
             uint256 executedQuote
         )
